+++ Network/API/Telegram/Bot/Object/Update/Message/Content/Info.hs
@@ -1,9 +1,10 @@
 module Network.API.Telegram.Bot.Object.Update.Message.Content.Info (Info (..), module Exports) where
 
 import Network.API.Telegram.Bot.Object.Update.Message.Content.Info.Location as Exports
+import Network.API.Telegram.Bot.Object.Update.Message.Content.Info.Venue as Exports
 
 import "aeson" Data.Aeson (FromJSON (parseJSON), withObject, (.:), (.:?))
-import "aeson" Data.Aeson.Types (Object, Parser, Value)
+import "aeson" Data.Aeson.Types (Object, Parser, Value (Object))
 import "base" Control.Applicative ((<*>), (<|>))
 import "base" Control.Monad ((>>=))
 import "base" Data.Function (($))
@@ -14,11 +15,11 @@ import "text" Data.Text (Text)
 
 data Info = Point Location
 	| Contact Text (Maybe Text) Text (Maybe Text)
-	| Venue Location Text Text (Maybe Text) (Maybe Text)
+	| Meeting Venue
 	deriving Show
 
 instance FromJSON Info where
-	parseJSON = withObject "Info" $ \v -> contact v <|> venue v <|> point v where
+	parseJSON = withObject "Info" $ \v -> contact v <|> meeting v <|> point v where
 
 		contact :: Object -> Parser Info
 		contact v = v .: "contact" >>= info where
@@ -28,13 +29,8 @@ instance FromJSON Info where
 				<$> i .: "first_name" <*> i .:? "last_name"
 				<*> i .: "phone_number" <*> i .:? "vcard"
 
-		venue :: Object -> Parser Info
-		venue v = v .: "venue" >>= info where
-
-			info :: Value -> Parser Info
-			info = withObject "Venue" $ \i -> Venue
-				<$> i .: "location" <*> i .: "title" <*> i .: "address"
-				<*> i .:? "foursquare_id" <*> i .:? "foursquare_type"
+		meeting :: Object -> Parser Info
+		meeting v = Meeting <$> parseJSON (Object v)
 
 		point :: Object -> Parser Info
 		point v = Point <$> v .: "location"